reports/pyproject.toml
[tool.poetry]
name = "reports"
version = "0.1.0"
description = "Bemo Financial Reports"
authors = ["Ruslan <rk@bemo.finance>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.8.2"
pydantic-settings = "^2.3.4"
httpx = "^0.27.0"
pandas = "^2.2.2"
loguru = "^0.7.2"
pytoniq = "^0.1.39"
python-dateutil = "^2.9.0.post0"
fastapi = "^0.111.1"
uvicorn = "^0.30.1"
sentry-sdk = {extras = ["loguru"], version = "^2.10.0"}


[tool.poetry.group.dev.dependencies]
ipython = "^8.26.0"
ruff = "^0.5.1"
isort = "^5.13.2"
mypy = "1.11.0"
black = "^24.4.2"
pre-commit = "^3.7.1"
dotenv-linter = "^0.5.0"
pandas-stubs = "^2.2.2.240603"
types-python-dateutil = "^2.9.0.20240316"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
profile = "black"
multi_line_output = 3

[tool.black]
exclude = '''
/(
  | .venv
)/
'''
line_length = 120


[tool.ruff]
lint.select = [
    "E", # pycodestyle
    "F", # Pyflakes
    "N", # pep8-naming
    "UP", # pyupgrade
    "BLE", # flake8-blind-except
    "B", # flake8-bugbear
    "A", # flake8-builtins
    "C4", # flake8-comprehensions
    "DTZ", # flake8-datetimez
    "DJ", # flake8-django
    "EXE", # flake8-executable
    "FA", # flake8-future-annotations
    "ISC", # flake8-implicit-str-concat
    "G", # flake8-logging-format
    "INP", # flake8-logging-format
    "PIE", # flake8-pie
    "T20", # flake8-print
    "PYI", # flake8-pyi
    "PT", # flake8-pyi
    "Q", # flake8-quotes
    "SIM", # flake8-simplify
    "TID", # flake8-tidy-imports
    "INT", # flake8-gettext
    "PTH", # flake8-use-pathlib
    "FIX", # flake8-fixme
    # Ruff-specific rules
    "RUF008", # Do not use mutable default values for dataclass attributes
    "RUF009", # Do not perform function call {name} in dataclass defaults
    "RUF010", # Use explicit conversion flag
    "B035", # Dictionary comprehension uses static key: {key}
    "RUF012", # Mutable class attributes should be annotated with typing.ClassVar
    "RUF013", # PEP 484 prohibits implicit Optional
    "RUF015", # Prefer next({iterable}) over single element slice
    "RUF016", # Slice in indexed access to type {value_type} uses type {index_type} instead of an integer.
    "RUF100", # Unused noqa directive
    "RUF200", # Failed to parse pyproject.toml
]
lint.ignore = [
    "FIX002",
    "N806", # Variable `UpperCase` in function should be lowercase
    "B008",
]

fix = true
line-length = 120
target-version = "py312"
extend-exclude = [".venv", "venv"]


[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unreachable = true
warn_no_return = true
namespace_packages = true
explicit_package_bases = true
disallow_untyped_defs = true
disallow_untyped_calls = false
strict = true
strict_equality = true
enable_incomplete_feature = ["NewGenericSyntax"]

[[tool.mypy.overrides]]
module = ["pytoniq"]
ignore_missing_imports = true
